requirements.txt
Pillow==9.5.0
Pillow==9.5.0
streamlit==1.21.0
